Integrate the code in the digest and cryptoutil modules directly into the sha1 module.
