+++ src/librustpkg/workcache_support.rs
@@ -12,8 +12,7 @@ use std::rt::io;
 use std::rt::io::extensions::ReaderUtil;
 use std::rt::io::file::FileInfo;
 use extra::workcache;
-use sha1::Sha1;
-use digest::Digest;
+use sha1::{Digest, Sha1};
 
 /// Hashes the file contents along with the last-modified time
 pub fn digest_file_with_date(path: &Path) -> ~str {
